Move API requests to requests.js and display number of likes
