+++ src/requests.js
@@ -1,0 +1,15 @@
+const getDogsData = async () => {
+  const url = 'https://api.thedogapi.com/v1/breeds?page=0&limit=9';
+  const response = await fetch(url, {
+    headers: { 'x-api-key': process.env.API_KEY },
+  });
+  return response.json();
+};
+
+const getLikes = async () => {
+  const url = `https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/${process.env.INVOLVEMENT_ID}/likes`;
+  const response = await fetch(url);
+  return response.json();
+};
+
+export { getDogsData, getLikes };
