+++ src/style.css
@@ -113,6 +113,10 @@ header {
   font-size: 40px;
 }
 
+.main-section-card p {
+  margin: 0;
+}
+
 .pop-image {
   width: 100%;
   height: 400px;
